requirements.txt
Flask>=2.2.2
requests==2.22.0
mysql-connector-python
